Move public Queue implementations out of internal
