+++ platform-sdk/consensus-concurrent/src/main/java/org/hiero/consensus/concurrent/framework/internal/ThreadBuildingUtils.java
@@ -5,6 +5,7 @@ import edu.umd.cs.findbugs.annotations.NonNull;
 import java.util.concurrent.BlockingQueue;
 import java.util.concurrent.LinkedBlockingQueue;
 import org.hiero.consensus.concurrent.framework.config.QueueThreadMetricsConfiguration;
+import org.hiero.consensus.concurrent.framework.queue.MeasuredBlockingQueue;
 
 /**
  * Static utility methods for the threading framework
