+++ platform-sdk/consensus-concurrent/src/main/java/org/hiero/consensus/concurrent/framework/queue/MeasuredBlockingQueue.java
@@ -1,5 +1,5 @@
 // SPDX-License-Identifier: Apache-2.0
-package org.hiero.consensus.concurrent.framework.internal;
+package org.hiero.consensus.concurrent.framework.queue;
 
 import com.swirlds.metrics.api.IntegerAccumulator;
 import com.swirlds.metrics.api.Metrics;
