+++ platform-sdk/consensus-gossip-impl/src/main/java/org/hiero/consensus/gossip/impl/gossip/sync/SyncMetrics.java
@@ -19,8 +19,8 @@ import java.util.Objects;
 import java.util.concurrent.BlockingQueue;
 import java.util.concurrent.ConcurrentHashMap;
 import java.util.concurrent.LinkedBlockingQueue;
-import org.hiero.consensus.concurrent.framework.internal.MeasuredBlockingQueue;
-import org.hiero.consensus.concurrent.framework.internal.MeasuredBlockingQueue.Config;
+import org.hiero.consensus.concurrent.framework.queue.MeasuredBlockingQueue;
+import org.hiero.consensus.concurrent.framework.queue.MeasuredBlockingQueue.Config;
 import org.hiero.consensus.gossip.impl.gossip.shadowgraph.ShadowgraphSynchronizer;
 import org.hiero.consensus.gossip.impl.gossip.shadowgraph.SyncPhase;
 import org.hiero.consensus.gossip.impl.gossip.shadowgraph.SyncResult;
